requirements.txt
polars
fastexcel
scikit-learn
psycopg2-binary
apache-airflow-providers-postgres
apache-airflow-providers-amazon
backoff
nltk
seaborn
lxml
